requirements.txt
Flask==2.3.3
scenedetect[opencv-headless]==0.6.2
opencv-python-headless==4.8.0.76
transformers==4.33.2
Pillow==10.0.1
torch==2.0.1
psycopg2==2.9.7
python-dotenv==1.0.0
numpy==1.26.0
sentence-transformers==2.2.2
av[no-binary]==10.0.0
huggingface_hub==0.17.3
deepface==0.0.79
openai==0.28.1
# retina-face==0.0.13
# keras==2.14.0